pyproject.toml
[project]
name = "labelweb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.0",
    "pillow>=12.1.0",
    "uvicorn>=0.40.0",
]
